lang: schema macro
