+++ crates/dojo-lang/src/lib.rs
@@ -5,6 +5,7 @@
 //! Learn more at [dojoengine.gg](http://dojoengine.gg).
 pub mod compiler;
 pub mod component;
+pub mod schema;
 pub mod inline_macros;
 mod manifest;
 pub mod plugin;
